Final-Project/Project-2-Agent/requirements.txt
# Requirements for Final Project 2: Agent
# pip install -r requirements.txt

langchain
langchain-openai
python-dotenv
pandas
ipykernel
